# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type AuthUserRole = typeof AuthUserRole[keyof typeof AuthUserRole] | null;


export const AuthUserRole = {
  manager: 'manager',
  inspector: 'inspector',
} as const;

export interface AuthUser {
  id: string;
  email?: string | null;
  firstName?: string | null;
  lastName?: string | null;
  profileImageUrl?: string | null;
  role?: AuthUserRole;
}

export interface AuthUserEnvelope {
  user: AuthUser | null;
}

export interface ExchangeMobileAuthorizationCodeBody {
  code: string;
  code_verifier: string;
  redirect_uri: string;
  state: string;
  nonce?: string | null;
}

export interface ExchangeMobileAuthorizationCodeResponse {
  token: string;
}

export interface LogoutMobileSessionResponse {
  success: boolean;
}

export type AuthorizationSessionHeaderParameter = string;

export type BeginBrowserLoginParams = {
returnTo?: string;
};