# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import * as zod from 'zod';


/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  "status": zod.string()
})


/**
 * @summary Get the currently authenticated user
 */
export const GetCurrentAuthUserHeader = zod.object({
  "Authorization": zod.string().optional()
})

export const GetCurrentAuthUserResponse = zod.object({
  "user": zod.union([zod.object({
  "id": zod.string(),
  "email": zod.string().nullish(),
  "firstName": zod.string().nullish(),
  "lastName": zod.string().nullish(),
  "profileImageUrl": zod.string().nullish(),
  "role": zod.enum(['manager', 'inspector']).nullish()
}),zod.null()])
})


/**
 * @summary Start the browser OIDC login flow
 */
export const BeginBrowserLoginQueryParams = zod.object({
  "returnTo": zod.coerce.string().optional()
})


/**
 * @summary Exchange mobile authorization code for a session token
 */
export const ExchangeMobileAuthorizationCodeBody = zod.object({
  "code": zod.string(),
  "code_verifier": zod.string(),
  "redirect_uri": zod.string(),
  "state": zod.string(),
  "nonce": zod.string().nullish()
})

export const ExchangeMobileAuthorizationCodeResponse = zod.object({
  "token": zod.string()
})


/**
 * @summary Log out mobile session
 */
export const LogoutMobileSessionResponse = zod.object({
  "success": zod.boolean()
})